backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
numpy
scikit-learn
python-dotenv
google-generativeai
